Extract FeaturesCollectionReference from GroundFirestore
